+++ src/main/java/insane96mcp/iguanatweaksexpanded/setup/client/ISEBookCategory.java
@@ -6,7 +6,6 @@ public enum ISEBookCategory implements StringRepresentable {
     FORGE_MISC("forge_misc"),
     BLAST_FURNACE_MISC("blast_furnace_misc"),
     SOUL_BLAST_FURNACE_MISC("soul_blast_furnace_misc"),
-    FLETCHING_MISC("fletching_misc"),
     UNKNOWN("unknown");
 
     public static final StringRepresentable.EnumCodec<ISEBookCategory> CODEC = StringRepresentable.fromEnum(ISEBookCategory::values);